Updated DAO Classes To Use Long as ID
+ Updated HibernateAdminDao to use Long for ID values.
+ Updated HibernateStudentDao to use Long for ID values.
+ Updated HibernateCourseDao to use Long for ID values.
+ Updated HibernateScJoinHolderDao to use Long for ID values.

+++ src/main/java/io/craigmiller160/school/repo/HibernateAdminDao.java
@@ -91,7 +91,7 @@ public class HibernateAdminDao implements GenericEntityDaoBean<Administrator> {
 	 * was set to null.
 	 */
 	@Override
-	public Administrator getEntityById(int adminId) {
+	public Administrator getEntityById(Long adminId) {
 		return (Administrator) sessionFactory.getCurrentSession()
 				.get(Administrator.class, adminId);
 	}
